+++ source/extensions/omni.replicator.isaac/docs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [1.1.0] - 2022-07-22
+### Added
+- Added mass, inertia, material properties, rest offset, and contact offset for rigid prim view randomization
+- Added mass, inertia, and tendon properties for articulation view randomization
+- Added additive and scaling operations for orientation randomization
+
 ## [1.0.1] - 2022-07-20
 ### Changed
 - Changed rigid_body_view to rigid_prim_view